move pruning to a new file
